UPDATE: In TaskRunner.createSymlinksInOutputDirectory(), now check before removing old link if the link is a symbolic link.

+++ src/main/java/fr/ens/transcriptome/eoulsan/core/workflow/TaskRunner.java
@@ -430,10 +430,14 @@ public class TaskRunner {
 
         try {
 
-          // TODO Use the Java 7 api to check if existing link is really a link
-          // Remove existing file/symlink
+          // Remove existing symlink
           if (link.exists()) {
-            link.delete();
+
+            if (link.getMetaData().isSymbolicLink()) {
+              link.delete();
+            } else {
+              throw new IOException();
+            }
           }
 
           // Create symbolic link
